+++ Modules/Microsoft365DSC/Examples/Resources/EXOPartnerApplication/2-Update.ps1
@@ -19,7 +19,6 @@ Configuration Example
         {
             Name                                = "HRApp"
             ApplicationIdentifier               = "00000006-0000-0dd1-ac00-000000000000"
-            AccountType                         = "OrganizationalAccount"
             Enabled                             = $False # Updated Property
             Ensure                              = "Present"
             Credential                          = $Credscredential