+++ Modules/Microsoft365DSC/Examples/Resources/EXOSafeLinksRule/1-Create.ps1
@@ -20,9 +20,9 @@ Configuration Example
             Identity                  = "Research Department URL Rule"
             Comments                  = "Applies to Research Department, except managers"
             Enabled                   = $True
-            ExceptIfSentToMemberOf    = "Research Department Managers"
+            ExceptIfSentToMemberOf    = "Executives"
             SafeLinksPolicy           = "Marketing Block URL"
-            SentToMemberOf            = "Research Department"
+            SentToMemberOf            = "Legal Team"
             Ensure                    = "Present"
             Credential                = $Credscredential
         }